add Imediator instead of services to controllers and make some changes to resolve errors. register every services in its own layer instead of rgister them one by one in program.cs

+++ Resume.Domain/ViewModels/Information/EditInformationViewModel.cs
@@ -3,7 +3,7 @@ using Resume.Domain.ViewModels.Common;
 
 namespace Resume.Domain.ViewModels.Information
 {
-    public class EditInformationViewModel : BaseViewModel<ulong>
+    public class EditInformationViewModel : BaseViewModel<int>
     {
         [Display(Name = "آواتار")]
         public string Avatar { get; set; }
